Add submission per question in quiz-solve:
Whenever a user answers a question, this data is saved in DB, so that if the user leaves the quiz half way and comes back to solving the same quiz later, they can continue from where they left off.
Also the back button is added for users to navigate to the previous questions.