Improve UX in terms of solving individual questions in quizzes:
Previously, the user was not able to navigate between the questions unless they provide an answer to the current question.

Now, the user can see any question they want without having to answer anything using the previous & next buttons. Also, if the user answers any of the questions, navigate to another question, and then come back to the initial question, they still view the answer they provided to the question.